Simplify install

+++ Vagrantfile
@@ -3,11 +3,18 @@
 
 require 'yaml'
 
+if !File.exist?( "files/vagrant.yml" )
+  FileUtils.cp( "files/vagrant.yml-EXAMPLE", "files/vagrant.yml" )
+end
 vagrantConfig = YAML::load_file( "files/vagrant.yml" )
 
 Vagrant.configure("2") do |config|
   CLOUD_CONFIG_PATH = "files/user-data"
   $use_aws = false
+
+  if !File.exist?( "files/fig.config" )
+      FileUtils.cp( "files/fig.config-EXAMPLE", "files/fig.config" )
+  end
 
   config.vm.provider :aws do |aws, override|
     $use_aws = true
@@ -68,13 +75,17 @@ Vagrant.configure("2") do |config|
       disabled: true
   end
 
-  FileUtils.cp( "files/auth.json", "dockerfiles/ezpublish/prepare" )
+  if File.exist?( "files/auth.json" )
+    FileUtils.cp( "files/auth.json", "dockerfiles/ezpublish/prepare" )
+  end
   FileUtils.cp( "files/etcd_functions", "dockerfiles/mysql" )
 
   # Install fig on vagrant machine
   config.vm.provision :shell, :inline => "
-    if [ ! -f /fig ]; then \
-      curl -L https://github.com/docker/fig/releases/download/1.0.0/fig-`uname -s`-`uname -m` > /fig; chmod +x /fig; \
+    if [ ! -f /opt/bin/fig ]; then \
+      mkdir -p /opt/bin
+      cp /vagrant/resources/fig /opt/bin
+      chmod +x /opt/bin/fig
     fi
   "
   if vagrantConfig['debug']['disable_docker_provision'] == false
